requirements.txt
prometheus_client>=0.20
requests>=2.32
PyYAML>=6.0.2
